# public/swagger.yaml
openapi: 3.0.0
info:
  title: API rute za personalizovane planere
  version: 1.0.0
  description: API za nasu prodavnicu personalizovanih planera
servers:
  - url: http://localhost:3000
paths:
...
  /api/auth/register:
    post:
      summary: Registracija korisnika
      description: |
        Kreira novog korisnika u bazi.
        Nakon uspesne registracije automatski generise auth token
        i postavlja HTTP-only cookie.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "200":
          description: Uspesna registracija
          headers:
            Set-Cookie:
              description: HTTP-only auth token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "400":
          description: Greska u validaciji ili email vec postoji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/admin/korisnik-uloga:
    put:
      summary: Promeni ulogu korisnika (admin only)
      description: |
        Omogucava adminu da promeni ulogu korisnika.
        Ako je korisnik admin - postaje ulogovani.
        Ako je korisnik ulogovani - postaje admin.
        Endpoint zahteva validan AUTH_COOKIE i admin privilegije.
      tags:
        - Korisnici
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userID
              properties:
                userID:
                  type: uuid
                  example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        "200":
          description: Uloga uspesno promenjena
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "400":
          description: Nedostaje userID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Nema userID
        "401":
          description: Nije autentifikovan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Nije autentifikovan
        "403":
          description: Zabranjen pristup (nije admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Zabranjen pristup korisniku
        "404":
          description: Korisnik ne postoji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Korisnik ne postoji
        "500":
          description: Greska servera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: AUTH_COOKIE
  schemas:
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - ime
        - prezime
        - adresa
        - password
      properties:
        username:
          type: string
          example: marko123
        email:
          type: string
          format: email
          example: marko@email.com
        ime:
          type: string
          example: Marko
        prezime:
          type: string
          example: Markovic
        adresa:
          type: string
          example: Bulevar 123
        password:
          type: string
          format: password
          example: mojaLozinka123
        uloga:
          type: string
          example: ulogovani
          description: Ignorise se u backendu (uvijek se postavlja na "ulogovani")
    UserResponse:
      type: object
      properties:
        id:
          type: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: marko123
        email:
          type: string
          example: marko@email.com
        ime:
          type: string
          example: Marko
        prezime:
          type: string
          example: Markovic
        adresa:
          type: string
          example: Bulevar 123
        uloga:
          type: string
          example: ulogovani
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Email postoji u bazu! 